Add a script to update and patch nixpkgs.
The script pulls nixpkgs from a git repository, pulls the latest
revision released on a given Nix channel, and merges in changes from a
list of channels.

The purpose of the script is to allow me to maintain an up-to-date copy
of nixpkgs which also contains certain changes I've made that have not
yet made their way into the official Nix channels.